v2_backend/requirements-v2.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
httpx
python-multipart
openpyxl
python-docx
pypdf
psycopg[binary]
